Updated movement system
Movement now works differently. Calling the Helirin::move() function now adds a value to the movement vector, which is processed once all the movements are taken into account. The main improvement is it allows us to see when stuff is moving diagonally, and thus if you collide into something while moving diagonally, you'll be pushed back in the correct distance.

Also lowered the cooldown on the collision reaction and increased the distance, to make it more snappy like the actual game.