Added Theme class
Big update this one. The main point was to create a 'Theme' class for the Level
one, so that it was easier to manage the various level themes. Along the way, I
ran into a couple more problems - namely that I suck at C++. Apparently you
can't actually define the values for stuff in header files. I was going to use
'extern', but I realised most of the values are only needed by the accompanying
source file anyway, so they've all just been moved.